requirements.txt
aerich==0.7.2
aiodns==3.2.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
Brotli==1.1.0
bs4==0.0.2
cffi==1.17.1
click==8.1.7
decorator==5.1.1
dictdiffer==0.9.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.114.2
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.9
ipython==8.27.0
iso8601==1.1.0
jedi==0.19.1
lxml==5.3.0
matplotlib-inline==0.1.7
multidict==6.1.0
parso==0.8.4
pexpect==4.9.0
playwright==1.47.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
pycares==4.4.0
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pyee==12.0.0
Pygments==2.18.0
pypika-tortoise==0.1.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
rapidfuzz==3.9.7
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.38.5
termcolor==2.4.0
toml==0.10.2
tomlkit==0.13.2
tortoise-orm==0.21.6
traitlets==5.14.3
typing_extensions==4.12.2
ua-generator==1.0.4
uvicorn==0.30.6
validators==0.34.0
wcwidth==0.2.13
yarl==1.11.1
